requirements.txt
streamlit
docxtpl
python-docx
pyyaml
